Flatten the fluent mappers into a single class per connector
The fluent mapper was a nested ReverseCompositeMapper of a path-provider mapper plus a
metadata mapper, even though both the segment and the metadata come from the same
FluentMappingRegistry. Replace that with one class per connector that subclasses the
connector's path-provider mapper and overrides the forward TryGetMapping to layer the
registry metadata (and, for OPC UA, the class-level type-self fallback) on top. Reverse
lookup is inherited from the base path mapper.

This removes the nested composite, the FluentMetadataMapper and OpcUaFluentMetadataMapper
types, and their no-op reverse. MqttPathProviderMapper.TryGetMapping and
OpcUaPathProviderMapper.TryGetMapping become virtual to allow the override. Behavior is
unchanged: forward merge order (metadata over path, property over type-self) and reverse
match the previous composite, verified by the connector unit and integration suites.
