+++ src/Namotion.Interceptor.Connectors.Tests/VerifyChecksTests.PublicApi.verified.txt
@@ -127,14 +127,6 @@ namespace Namotion.Interceptor.Connectors.Mapping
         public bool TryGetTypeMetadata(System.Type subjectType, string member, [System.Diagnostics.CodeAnalysis.NotNullWhen(true)] out TMetadata? metadata) { }
         public bool TryGetTypeSelfMetadata(System.Type type, [System.Diagnostics.CodeAnalysis.NotNullWhen(true)] out TMetadata? metadata) { }
     }
-    public class FluentMetadataMapper<TMetadata, TKey> : Namotion.Interceptor.Connectors.Mapping.IPropertyMapper<TMetadata>, Namotion.Interceptor.Connectors.Mapping.IReversePropertyMapper<TMetadata, TKey>
-        where TMetadata :  class, Namotion.Interceptor.Connectors.Mapping.IPropertyMapping<TMetadata>
-    {
-        public FluentMetadataMapper(Namotion.Interceptor.Connectors.Mapping.FluentMappingRegistry<TMetadata> registry) { }
-        protected Namotion.Interceptor.Connectors.Mapping.FluentMappingRegistry<TMetadata> Registry { get; }
-        public virtual bool TryGetMapping(Namotion.Interceptor.Registry.Abstractions.RegisteredSubjectProperty property, Namotion.Interceptor.IInterceptorSubject rootSubject, [System.Diagnostics.CodeAnalysis.NotNullWhen(true)] out TMetadata? mapping) { }
-        public System.Threading.Tasks.ValueTask<Namotion.Interceptor.Registry.Abstractions.RegisteredSubjectProperty?> TryGetPropertyAsync(TKey key, Namotion.Interceptor.Registry.Abstractions.RegisteredSubject subject, System.Threading.CancellationToken cancellationToken) { }
-    }
     public sealed class FluentPathProvider<TMetadata> : Namotion.Interceptor.Registry.Paths.PathProviderBase
         where TMetadata :  class
     {
